# docs/openapi.yaml
openapi: 3.1.0
info: {title: PPT Agent P8 API, version: 1.0.0-step1}
components:
  parameters:
    TaskId: {in: path, name: task_id, required: true, schema: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$'}}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: object, required: [code, message, diagnostic_id], properties: {code: {type: string}, message: {type: string}, diagnostic_id: {type: string}, agent_audit_id: {type: string}}}
...
    AgentAudits: {type: object, required: [audits], properties: {audits: {type: array, items: {type: object}}}}
...
    TaskState:
      type: object
      required: [task_id, stage, status, mode, sample_confirmed, blockers_resolved, delivery_confirmed, revision, waiting_reason, required_action]
      properties: {task_id: {type: string}, stage: {enum: [created, clarification, narrative, outline, sample, deck, review, delivery]}, status: {enum: [ready, running, waiting_for_user, paused, cancelled, failed, completed]}, mode: {enum: [manual, auto]}, sample_confirmed: {type: boolean}, blockers_resolved: {type: boolean}, delivery_confirmed: {type: boolean}, revision: {type: integer}, waiting_reason: {type: [string, 'null']}, required_action: {type: [string, 'null']}}
    Command:
      type: object
      additionalProperties: false
      required: [command_id, action]
      properties: {command_id: {type: string}, action: {type: string}, actor: {enum: [system, user]}, payload: {type: object}}
    Event:
      type: object
      additionalProperties: false
      required: [event_id, command_id, action, actor, request_hash, at, from, to, result]
      properties:
        event_id: {type: string}
        command_id: {type: string}
        action: {type: string}
        actor: {enum: [system, user]}
        request_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
        at: {type: string, format: date-time}
        from: {$ref: '#/components/schemas/TaskState'}
        to: {$ref: '#/components/schemas/TaskState'}
        result: {$ref: '#/components/schemas/TaskState'}
    Events: {type: object, required: [events], properties: {events: {type: array, items: {$ref: '#/components/schemas/Event'}}}}
    Version: {type: object, required: [kind, hash, metadata], properties: {kind: {type: string}, hash: {type: string, pattern: '^[0-9a-f]{64}$'}, metadata: {type: object}}}
    Versions: {type: object, required: [versions], properties: {versions: {type: array, items: {$ref: '#/components/schemas/Version'}}}}
    VersionContent: {type: object, required: [hash, content], properties: {hash: {type: string}, content: {type: string}}}
    Comparison: {type: object, required: [left, right, equal], properties: {left: {type: string}, right: {type: string}, equal: {type: boolean}}}
    InputView: {type: object, required: [state, snapshot], properties: {state: {$ref: '#/components/schemas/TaskState'}, snapshot: {type: [object, 'null']}, snapshot_hash: {type: string}, task_card: {type: object}, manifest: {type: object}, clarification: {type: object}, clarification_hash: {type: string}}}
    ClarificationResult: {type: object, required: [state, clarification_hash, questions, answers, invalidated, confirmed], properties: {state: {$ref: '#/components/schemas/TaskState'}, clarification_hash: {type: string}, questions: {type: array}, answers: {type: object}, invalidated: {type: array}, confirmed: {type: boolean}}}
...
  responses:
    Error: {description: Unified domain error, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}, example: {error: {code: validation_error, message: 请求字段无效, diagnostic_id: abc123}}}}}
paths:
...
  /v1/tasks/{task_id}/shell:
    get:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Server-authoritative application shell state and eight-stage navigation, content: {application/json: {schema: {type: object}, example: {task: {task_id: demo, stage: created, status: ready, mode: manual}, summary: {progress: 5}, stages: [{id: created, label: 任务/资料, status: current, lock_reason: null, href: '/tasks/demo?stage=created'}], active_jobs: []}}}}, '404': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/previews/{hash}:
    get:
      description: Read-only sandbox source for a sample or deck version owned by the task. The response forbids scripts, is not cached, and can only be framed by the same origin.
      parameters:
        - {$ref: '#/components/parameters/TaskId'}
        - {in: path, name: hash, required: true, schema: {type: string, pattern: '^[0-9a-f]{64}$'}}
      responses:
        '200': {description: Sanitized and previously validated HTML artifact, headers: {Cache-Control: {schema: {type: string}, example: 'private, no-store'}, X-Frame-Options: {schema: {type: string}, example: SAMEORIGIN}}, content: {text/html: {schema: {type: string}, example: '<!doctype html><html><body><section class="slide" id="slide-1"></section></body></html>'}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
  /v1/tasks/{task_id}/inspection:
    get:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Current hash-bound report, dispositions and delivery gate, content: {application/json: {schema: {type: object}, example: {report: null, dispositions: [], delivery_allowed: false}}}}, '404': {$ref: '#/components/responses/Error'}}
...
  /v1/tasks/{task_id}/inspection/mode:
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [mode], properties: {mode: {enum: [manual, auto]}}}}}}
      responses: {'200': {description: Mode changed for the next action, content: {application/json: {schema: {type: object}, example: {state: {mode: auto}}}}}, '400': {$ref: '#/components/responses/Error'}}
...
  /v1/tasks/{task_id}/samples:
    get:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Current selection, sample, confirmation and immutable history, content: {application/json: {schema: {type: object}, example: {state: {}, selection: null, sample: null, confirmation: null, versions: []}}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}
...
  /v1/tasks/{task_id}/deck:
    get:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Current deck plus immutable timeline with source and operator, content: {application/json: {schema: {type: object}, example: {state: {}, deck: null, versions: []}}}}, '404': {$ref: '#/components/responses/Error'}}
...
  /v1/tasks:
    get:
      responses: {'200': {description: Recently updated task states for the application shell, content: {application/json: {schema: {type: object, required: [tasks], properties: {tasks: {type: array, items: {$ref: '#/components/schemas/TaskState'}}}}, example: {tasks: [{task_id: demo, stage: created, status: ready, mode: manual, sample_confirmed: false, blockers_resolved: false, delivery_confirmed: false, revision: 0, waiting_reason: null, required_action: null}]}}}}, '500': {$ref: '#/components/responses/Error'}}
    post:
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [task_id], properties: {task_id: {type: string}, mode: {enum: [manual, auto]}}}, example: {task_id: demo, mode: auto}}}}
      responses: {'201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/TaskState'}, example: &task {task_id: demo, stage: created, status: ready, mode: auto, sample_confirmed: false, blockers_resolved: false, delivery_confirmed: false, revision: 0, waiting_reason: null, required_action: null}}}}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}:
    get:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Current state, content: {application/json: {schema: {$ref: '#/components/schemas/TaskState'}, example: *task}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/agent-audits:
    get:
      parameters:
        - {$ref: '#/components/parameters/TaskId'}
        - {in: query, name: job_id, schema: {type: string, pattern: '^job_[A-Za-z0-9_-]+$'}}
      responses: {'200': {description: Secret-free Agent audits correlated to this task and optional Job, content: {application/json: {schema: {$ref: '#/components/schemas/AgentAudits'}, example: {audits: []}}}}, '404': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/actions:
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Command'}, example: {command_id: cmd-1, action: advance, actor: system}}}}
      responses: {'200': {description: Stable command result, content: {application/json: {schema: {$ref: '#/components/schemas/TaskState'}, example: *task}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/input:
    get:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Frozen input, content: {application/json: {schema: {$ref: '#/components/schemas/InputView'}, example: {state: *task, snapshot: null}}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [source], properties: {source: {type: [object, string]}, source_format: {enum: [json, markdown]}, rebuild: {type: boolean}}}, example: {source: {goal: 产品介绍, audience: 客户, topic: 新品}, source_format: json, rebuild: false}}}}
      responses: {'200': {description: Normalized and frozen input, content: {application/json: {schema: {$ref: '#/components/schemas/InputView'}, example: {state: *task, snapshot: {snapshot_id: snapshot-1}, snapshot_hash: abc, task_card: {}, manifest: {}, clarification: {}, clarification_hash: def}}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/clarifications/{question_id}/answer:
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}, {in: path, name: question_id, required: true, schema: {type: string}}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [option], properties: {option: {type: string}, other: {type: string}}}, example: {option: Other, other: 管理层}}}}
      responses: {'200': {description: Answer saved as a new clarification version, content: {application/json: {schema: {$ref: '#/components/schemas/ClarificationResult'}, example: {state: *task, clarification_hash: abc, questions: [], answers: {missing-audience: 管理层}, invalidated: [], confirmed: true}}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/planning:
    get:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Current narrative, outline and visible version history, content: {application/json: {schema: {type: object}, example: {state: {}, narrative: null, outline: null, versions: []}}}}, '400': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/narrative:
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Authoritative narrative edit saved as a new version, content: {application/json: {schema: {type: object}, example: {state: {}, narrative: {}, outline: null, versions: []}}}}, '400': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/narrative/generate:
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Narrative generated with a traceable skill slice, content: {application/json: {schema: {type: object}, example: {state: {}, narrative: {}, outline: null, versions: []}}}}, '400': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/outline:
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Validated authoritative outline edit and affected slides, content: {application/json: {schema: {type: object}, example: {state: {}, narrative: {}, outline: {}, versions: []}}}}, '400': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/outline/generate:
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Outline generated from current narrative and frozen resources, content: {application/json: {schema: {type: object}, example: {state: {}, narrative: {}, outline: {}, versions: []}}}}, '400': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/planning/rollback:
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Historical content copied into a new current version, content: {application/json: {schema: {type: object}, example: {state: {}, narrative: {}, outline: {}, versions: []}}}}, '400': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/events:
    get:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Audit event list, content: {application/json: {schema: {$ref: '#/components/schemas/Events'}, example: {events: [{event_id: evt-1, command_id: cmd-1, action: advance, actor: system, request_hash: '0000000000000000000000000000000000000000000000000000000000000000', at: '2026-08-11T00:00:00+00:00', from: *task, to: *task, result: *task}]}}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/versions:
    get:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Version list, content: {application/json: {schema: {$ref: '#/components/schemas/Versions'}, example: {versions: [{kind: deck, hash: '0000000000000000000000000000000000000000000000000000000000000000', metadata: {source: fake}}]}}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/versions/{hash}:
    get:
      parameters: [{$ref: '#/components/parameters/TaskId'}, {in: path, name: hash, required: true, schema: {type: string, pattern: '^[0-9a-f]{64}$'}}]
      responses: {'200': {description: Immutable version, content: {application/json: {schema: {$ref: '#/components/schemas/VersionContent'}, example: {hash: '0000000000000000000000000000000000000000000000000000000000000000', content: '<html></html>'}}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/versions/compare:
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [left, right], properties: {left: {type: string}, right: {type: string}}}, example: {left: 0000000000000000000000000000000000000000000000000000000000000000, right: 0000000000000000000000000000000000000000000000000000000000000000}}}}
      responses: {'200': {description: Comparison result, content: {application/json: {schema: {$ref: '#/components/schemas/Comparison'}, example: {left: '0000000000000000000000000000000000000000000000000000000000000000', right: '0000000000000000000000000000000000000000000000000000000000000000', equal: true}}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/summary:
    get:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      responses: {'200': {description: Orchestrator-safe status summary without dialogue or reasoning, content: {application/json: {schema: {type: object}, example: {task_id: demo, stage: delivery, progress: 100, status: completed}}}}, '404': {$ref: '#/components/responses/Error'}}
...
  /v1/tasks/{task_id}/issues/{issue_id}/disposition:
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}, {in: path, name: issue_id, required: true, schema: {type: string}}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [action], properties: {action: {enum: [agent_fix, manual, waive, defer]}, rationale: {type: string}, actor: {enum: [user, system]}}}, example: {action: waive, rationale: 用户接受该版式风险, actor: user}}}}
      responses: {'200': {description: Version-bound audited disposition, content: {application/json: {schema: {type: object}, example: {disposition_hash: abc, delivery_allowed: true}}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /v1/tasks/{task_id}/issues/dispositions/batch:
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [issue_ids, action], properties: {issue_ids: {type: array, minItems: 1, uniqueItems: true, items: {type: string}}, action: {enum: [agent_fix, manual, waive, defer]}, rationale: {type: string}}}}}}
      responses: {'200': {description: Explicit same-action batch disposition with returned batch scope, content: {application/json: {schema: {type: object}, example: {batch_scope: [issue-1], disposition_hashes: [abc]}}}}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}